requirements.txt
aiohttp==3.11.13
beautifulsoup4==4.13.3
celery==5.4.0
croniter==6.0.0
Django==5.1.7
django-allauth[socialaccount,mfa]==65.4.1
django-celery-beat==2.7.0
django-celery-results==2.5.1
django_debug_toolbar==5.0.1
django-health-check==3.18.3
django-model-utils==5.0.0
django-select2==8.3.0
django-simple-history==3.8.0
django-widget-tweaks==1.5.0
fakeredis==2.27.0
gunicorn==23.0.0
Pillow==11.1.0
psycopg[binary]==3.2.5
python-decouple==3.8
redis[hiredis]==5.2.1
requests==2.32.3
requests-ratelimiter==0.7.0
unidecode==1.3.8
